+++ CHANGELOG.md
@@ -12,12 +12,12 @@
   - [KAFKA-207](https://jira.mongodb.org/browse/KAFKA-207) Improved efficiency of heartbeats by making them tombstone messages.
   - [KAFKA-174](https://jira.mongodb.org/browse/KAFKA-174) Improved error messages when using invalid pipeline operators.
   - [KAFKA-194](https://jira.mongodb.org/browse/KAFKA-194) Added support for Qlik Replicate CDC.
-  - [KAFKA-209](https://jira.mongodb.org/browse/KAFKA-209) Log a warning message when there the `_id` value and the
-    id strategy is configured not to overwrite the existing `_id` value.
 
 ### Bug Fixes
   - [KAFKA-195](https://jira.mongodb.org/browse/KAFKA-195) Fixed topics.regex sink validation issue for synthetic config property
   - [KAFKA-203](https://jira.mongodb.org/browse/KAFKA-203) Fixed sink NPE issue when using with confluent connect 6.1.0
+  - [KAFKA-209](https://jira.mongodb.org/browse/KAFKA-209) Fixed `_id` always being projected even if not explicitly allowed or blocked.
+    Log a warning message when there the `_id` value and the id strategy is configured not to overwrite the `_id`.
 
 ## 1.4.0